Move MemRefType et al. to builtin.
